Update FileWriter to write multiple files for differeing instances of the same UUID.

+++ pipeline/io/arches.py
@@ -1,21 +1,32 @@
-from bonobo.config import Configurable, Option
-import requests
 import os
+import os.path
+import hashlib
 
+import requests
+from bonobo.config import Configurable, Option
+from pipeline.util import ExclusiveValue
 
 class FileWriter(Configurable):
 	directory = Option(default="output")
 
 	def __call__(self, data: dict):
 		d = data['_OUTPUT']
+		uuid = data['uuid']
 		dr = os.path.join(self.directory, data['_ARCHES_MODEL'])
-		if not os.path.exists(dr):
-			os.mkdir(dr)
-		fn = os.path.join(dr, "%s.json" % data['uuid'])
-		fh = open(fn, 'w')
-		fh.write(d)
-		fh.close()
-		return data
+		with ExclusiveValue(dr):
+			if not os.path.exists(dr):
+				os.mkdir(dr)
+		ddr = os.path.join(dr, uuid)
+		with ExclusiveValue(ddr):
+			if not os.path.exists(ddr):
+				os.mkdir(ddr)
+			h = hashlib.md5(d.encode('utf-8')).hexdigest()
+			fn = os.path.join(ddr, "%s.json" % (h,))
+			if not os.path.exists(fn):
+				fh = open(fn, 'w')
+				fh.write(d)
+				fh.close()
+			return data
 
 class ArchesWriter(Configurable):
 	endpoint = Option(default="http://localhost:8001/resources/")
